pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hackaikit"
version = "1.0.0"
description = "A comprehensive AI/ML toolkit for rapid prototyping and hackathons"
readme = "README.md"
authors = [
    {name = "HackAI-Kit Team"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.8"
dependencies = [
    # Core dependencies
    "numpy>=1.20.0",
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
    "fastapi>=0.95.0",
    "uvicorn>=0.22.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
    
    # External API clients
    "openai>=1.0.0",
    "google-generativeai>=0.3.0",
    "huggingface-hub>=0.19.0",
    
    # Data visualization
    "matplotlib>=3.5.0",
    "seaborn>=0.12.0",
]

[project.optional-dependencies]
# Module-specific dependencies
supervised = [
    "xgboost>=1.5.0",
]
unsupervised = [
    "hdbscan>=0.8.0",
    "umap-learn>=0.5.0",
    "sentence-transformers>=2.2.0",
]
nlp = [
    "transformers>=4.30.0",
    "tokenizers>=0.14.0",
    "nltk>=3.8.0",
    "spacy>=3.6.0",
]
computer_vision = [
    "opencv-python>=4.7.0",
    "pillow>=9.4.0",
    "ultralytics>=8.0.0",  # YOLOv8
    "torch>=2.0.0",
    "torchvision>=0.15.0",
]
time_series = [
    "prophet>=1.1.0",
    "statsmodels>=0.14.0",
]
reinforcement_learning = [
    "gymnasium>=0.28.0",  # New version of Gym
    "stable-baselines3>=2.0.0",
]
deep_learning = [
    "torch>=2.0.0",
    "tensorflow>=2.12.0",
]
sensor_signal = [
    "scipy>=1.10.0",
]
tinyml = [
    "tensorflow>=2.12.0",
    "tensorflow-lite>=2.12.0",
]
dev = [
    "pytest>=7.3.1",
    "black>=23.3.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.3.0",
]
docs = [
    "mkdocs>=1.4.3",
    "mkdocs-material>=9.1.15",
]

# Define all modules for easy installation
all = [
    "hackaikit[supervised]",
    "hackaikit[unsupervised]",
    "hackaikit[nlp]",
    "hackaikit[computer_vision]",
    "hackaikit[time_series]",
    "hackaikit[reinforcement_learning]",
    "hackaikit[deep_learning]",
    "hackaikit[sensor_signal]",
    "hackaikit[tinyml]",
]

[project.scripts]
hackaikit-api = "hackaikit.api.main:run_api"

[project.urls]
Homepage = "https://github.com/user/hackaikit"
Documentation = "https://hackaikit.readthedocs.io"
Issues = "https://github.com/user/hackaikit/issues"

[tool.setuptools]
packages = ["hackaikit"]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310"]

[tool.isort]
profile = "black"
line_length = 88 